Move new LateParsing resolution related logic to SemaBoundsSafety.cpp
